requirements.txt
cryptography==2.8
inflection==0.3.1
mypy==0.740
pylint==2.4.4
PyMySQL==0.9.3
psycopg2-binary==2.8.4
pytest==5.2.2
pytest-cov==2.8.1
pytest-mock==1.12.1
pytest-mypy==0.4.2
pytest-sugar==0.9.2